Move MarcReader to separate namespace, split serialization functions out of it.
